requirements.txt
flask
geopandas
psycopg2-binary
mangum
dotenv
boto3